% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.kfold.R
\name{accuracy.kfold}
\alias{accuracy.kfold}
\title{Estimate assignment accuracies of K-fold cross-validation results}
\usage{
accuracy.kfold(dir = NULL)
}
\arguments{
\item{dir}{A character string to specify the folder that has your K-fold cross-validation results. A slash should be included at the end (e.g., dir="YourFolderName/").}
}
\value{
This function outputs the results in a text file (a table). It can return a data frame when a returning object is specified.
}
\description{
This function allows you to estimate assignment accuracies of K-fold cross-validation results. The output results can be used to make assignment accuracy plots (use function accuracy.plot) and membership probability plot (use function membership.plot)
}
\examples{
# results <- accuracy.kfold(dir="YourFolderName/")
}

