% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.MC.R
\name{accuracy.MC}
\alias{accuracy.MC}
\title{Estimate assignment accuracies of Monte-Carlo cross-validation results}
\usage{
accuracy.MC(dir = NULL)
}
\arguments{
\item{dir}{A character string to specify the folder that has your Monte-Carlo cross-validation results. A slash should be included at the end (e.g., dir="YourFolderName/").}
}
\value{
This function outputs the results in a text file (a table). It can return a data frame when a returning object is specified.
}
\description{
This function allows you to estimate assignment accuracies of Monte-Carlo cross-validation results. The output results can be used to make assignment accuracy plots (use function accuracy.plot).
}

