% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.matrix.R
\name{assign.matrix}
\alias{assign.matrix}
\title{Make an assignment maxtrix from cross-validation results}
\usage{
assign.matrix(dir = NULL)
}
\arguments{
\item{dir}{A character string to specify the folder that has your cross-validation assignment results.}
}
\value{
The function returns a matrix in R console as well as a file named "assignment_matrix.txt" in the folder.
}
\description{
This function generates a pairwise assignment matrix with mean and variation of assignment accuracies estimated across all assignment tests.
}

