% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.MC.R
\name{assign.MC}
\alias{assign.MC}
\title{Population assignment test using Monte-Carlo cross-validation}
\usage{
assign.MC(x, train.inds = c(0.5, 0.7, 0.9), train.loci = c(0.1, 0.25, 0.5,
  1), loci.sample = "fst", iterations = 20, dir = NULL,
  pca.method = "kaiser-guttman", pca.loadings = F, model = "svm",
  svm.kernel = "linear", svm.cost = 1, ntree = 50, processors = 999,
  ...)
}
\arguments{
\item{x}{An input object which should be the object (list) returned from the function read.genpop(), reduce.allele(), or compile.data().}

\item{train.inds}{The number (integer greater than 1) or proportion (float between 0 and 1) of individuals (observations) from each population to be used as training data. Use a vector to specify multiple sets of training individuals. No mixture of integer and float in the vector.}

\item{train.loci}{The proportion (float between 0 and 1) of loci to be used as training data. Use a vector to specify multiple sets of training loci.}

\item{loci.sample}{Locus sampling method, "fst" or "random". If loci.sample="fst" (default), training sample's locus Fst will be estimated and prioritized. The loci will be sampled based on the Fst values. For example, when train.loci=0.1, the top 10 percents high Fst loci are sampled as training loci. If loci.sample="random", then random 10 percents of loci will be sampled as training loci.}

\item{iterations}{Resampling times (an integer) for each combination of training individuals and loci.}

\item{dir}{A character string to specify the name of output directory. A slash at the end must be included (e.g., dir="YourFolderName/")}

\item{pca.method}{A criterion to retain number of PCs. By default, it used Kaiser-Guttman criterion that the PC has eigenvalue greater than 1 will be retained as the new variable/feature. Users can set an integer to specify the number of PCs to be retained.}

\item{pca.loadings}{A logical variable (False or True) to determine whether it prints the loadings of training data to output text files.}

\item{model}{A character string to specify which method to use for creating predicting models. The options include "lda", "svm", "naiveBayes", "tree", and "randomForest".}

\item{svm.kernel}{A character string to specify which kernel to be used when using "svm" method.}

\item{svm.cost}{A number to specify the cost for "svm" method.}

\item{ntree}{A integer to specify how many trees to build when using "randomForest" method.}

\item{processors}{The number of processors to be used for parallel running. By default, it uses N-1 processors in your computer.}
}
\value{
You don't need to specify a name for the returned object when using this function. It automatically output results in text files to your designated folder.
}
\description{
This function uses Monte-Carlo cross-validation to perform assignment tests. It takes the object returned from the function read.genpop(), reduce.allele(), or compile.data() as input file, and saves output in text files. It is recommended to specify a folder name using the argument "dir" for saving output files. See argument description for more details.
}

