% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.rate.kfold.R
\name{assign.rate.kfold}
\alias{assign.rate.kfold}
\title{Estimate correct assignment rate restuls of K-fold cross-validation}
\usage{
assign.rate.kfold(dir = NULL)
}
\arguments{
\item{dir}{A character string to specify the folder that has your K-fold cross-validation results. A slash should be entered at the end.}
}
\value{
This function outputs the results in a text file (a table). If a returning object is specify, it returns the results in a data frame.
}
\description{
This function allows you to estimate correct assignment rates of K-fold cross-validation results. The output results can be used to make correct assignment rate plots (use function car.plot) and membership probability plot (use function prob.plot)
}
\examples{
results <- assign.rate.MC(dir="YourFolderName/")
}

