% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car.plot.R
\name{car.plot}
\alias{car.plot}
\title{Make a boxplot of correct assignment rate results from cross-validation (ggplot2 style)}
\usage{
car.plot(df, pop = "all")
}
\arguments{
\item{df}{A dataframe of your correct assignment rate results. It can be the object returned from the function assign.rate.MC() or assign.rate.kfold() or a data frame that is later imported using the function such as read.table().}

\item{pop}{A population name for making its plot. By default, it uses "all", meaning overall correct assignment rates. If a population name is specified, it only creates the plot for that population. The specified population name should match what you entered in read.genpop() earlier.}
}
\value{
This function returns a boxplot plot using the ggplot2 library. Users can modified (e.g., change color, text, etc.) the plot using functions provided by ggplot2 library.
}
\description{
This functions allows you to make a boxplot of correct assignment rate results estimated from Monte-Carlo or K-fold cross-validation.
}
\examples{
Your_df <- read.table("YourFolderName/Rate_of_N_tests_M_pops.txt", header=T)
car.plot(Your_df, pop="all")
}

