% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.loci.R
\name{check.loci}
\alias{check.loci}
\title{Check which loci frequently have high Fst across training sets}
\usage{
check.loci(dir = NULL, top.loci = 20)
}
\arguments{
\item{dir}{A character string to specify the folder with your cross-validation results. A slash should be entered at the end.}

\item{top.loci}{An integer to specify how many top informative loci to output.}
}
\value{
This function output the results in a text file. It includes the top N informative loci in N rows, and each row has a list of loci sorted by its occurrence.
}
\description{
This function reads through training locus file for each assignment test and counts the frequency of those loci and outputs the results in a text file.
}
\examples{
check.loci(dir="YourFolderName/", top.loci=20)
}

