% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.structure.R
\name{read.structure}
\alias{read.structure}
\title{Read Structure format file}
\usage{
read.structure(x, haploid = FALSE)
}
\arguments{
\item{x}{STRUCTURE file or path to the file. The filename extension (e.g., .txt) should be included.}

\item{haploid}{A logical variable (TRUE or FALSE) to specify whether your dataset is haploid data. Default is FALSE.}
}
\value{
This function returns a list comprising three elements. 1. YOU_NAME_IT$DataMatrix: A matrix of genetic data with a population name label ($popNameVector) in the last column. 2. YOU_NAME_IT$SampleID: A vector of sample ID. 3. YOU_NAME_IT$LocusName: A vector of locus name.
}
\description{
This function allows you to import a Structure format file into R.
}

