% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/membership.plot.R
\name{membership.plot}
\alias{membership.plot}
\title{Make a membership probability plot using results from K-fold cross-validation (ggplot2 style)}
\usage{
membership.plot(dir = NULL, style = NULL)
}
\arguments{
\item{dir}{A character string to specify the folder that has your K-fold cross-validation assignment results. A slash should be entered at the end.}

\item{style}{An option for output style. If style=1, it creates the plot which individuals on the x-axis are sorted by default/random. If style=2, individuals are sorted by the probabilities. If style=3, individuals are separated by each fold. If style=4, individuals are separated by fold and sorted by probability.}
}
\value{
his function returns a stacked-bar plot using the ggplot2 library. Users can modified the plot (e.g., change color, text, etc.) using functions provided by ggplot2 library.
}
\description{
This function allows you to make a membership probability plot (stacked-bar plot) using results estimated from K-fold cross-validation.
}
\examples{
# membership.plot(dir="YourFolderName/")
#If style is not specified, the program will ask you to enter after executing the function.
}

