% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.structure.R
\name{read.structure}
\alias{read.structure}
\title{Read Structure format file}
\usage{
read.structure(x, pop.names = NULL, haploid = FALSE)
}
\description{
This function allows you to import a Structure format file into R.
}

